Overhauled the changepoint functions
segment_finder replaces segments_finder. It is completely generalisable,
allowing all the columns to be specified. Allows auto sigma, global
sigma, or a list of sigma values as input. Allows a single trajectory to
be analysed from a table with or without a trajectory column, or all
trajectories to be analysed with a supplied trajectory column. Saves
output file names according to the analysis performed.